+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 删除 API `/api/broadcast/channels` | Delete API `/api/broadcast/channels`
 - 添加 API `/api/broadcast/getChannels` | Delete API `/api/broadcast/getChannels`
 - 添加 API `/api/network/echo` | Delete API `/api/network/echo`
+- 调整 API `/api/storage/getRecentDocs` 的 JSON Schema | Adjust the JSON Schema of API `/api/storage/getRecentDocs`
 
 ## v0.3.4 / 2023-11-15